fix #10058
The problem has been investigated. It consists of two parts:

1. Bad logic in ess-app.js, the original code uses

response.data.indexOf('0LOKZECwqdNGvKZFy3uB') !== -1
to determine whether the http request is a timeout request. However, it does not consider the API calls in parallel. Therefore, the multiple popup windows occurs.

2. Apache shiro store the redirection request, but it has no functionality to distinguish API calls and normal page request due to the AngualJS refreshing page uses sending page from back-end to front-end. Therefore, API calls may store as a redirection page for next time login.

Fix for #1:
I introduces a local-storage variable to indicate whether two API calls are in parallel and only allows one of them (randomly one) to pop up the alert.

Fix for #2:

The problem is just like Kevin said, i manually changed redirectURL to "/" (root path in logic) when savedRequest is an API call instead of any real URL.