Fix modal overflow issue, change large supply img
As part of the large supply image button commit:
2c409c0099bd762fbfbd7a1d130efffe77ecb12b , modals were set as overflow:
visible instead of overflow:auto.  This caused some side effects where
modal content would be cut off for certain modals on small screens.
This commit sets the modal default back to overflow:auto and implements
 the large supply image close button in a different way.  Large Supply
images now have a description pane with an 'X' icon to close in the
right corner.