Fix Time Record cache polling query
The time record cache polling query that polls for updates is not
working correctly.  It does not pick up time record updates for time
records that do not yet have entries.  This is due to a JOIN between the
record and entry tables.  This commit changes that to a LEFT JOIN to
enable the query to pick up empty time records.