Modify employee dao getLastUpdateTime query
Modify the getLastUpdateTime query to more resemble the employee query.
This query was returning last update times that were different from the
latest update times retrieved from getUpdatedEmployees.  This was most
likely due to CDSTATUS checks that were present in one query and not in
the other.  The getLastUpdateTime query now uses the same tables and
joins as the other employee queries (including getUpdatedEmployees) to
prevent differing update times going forward.

Fixes #10984