Create new api for time record review
It doesn't make sense to use the save time record api to review time
records.  Reviewing entails two things: action (approval/disapproval)
and remarks.  Posting and validating the fields of an entire time
record is overkill in this situation.

This commit adds a time record review api
that will make the appropriate modifications to a time record based on
posted review parameters.  This api is integrated into the time record
manage page controller.