Fix sup emp group issue from same-day transactions
If an employee has two different supervisor transactions on the same
day, they are supposed to be applied in the order specified by the
creation date and change number of the two different records.  This was
not occurring properly in the supervisor employee query, because the
table names were not specified for the columns in the ordering.
Also when this occurs, it will produce an emp. sup. info record with
 an empty effective date range.  These issues are addressed in this
commit.
 - Specify table names in TRANS_RANK order by in supervisor emp. group
   query
 - Prevent emp. sup. info records with empty effective date ranges from
   being added to a supervisor emp group in the SqlSupervisorDao

Refs #10989