Remove option to view attendance info for senators
Senators do not enter time or use accruals, so they should not be
selectable in the attendance management pages.  Additionally, some
senators are their own supervisor.  It does not make sense to allow
these senators to view their own data in the management section.  This
commit adds restrictions for employee selection in the management pages.
 - Add "supervisor" flag to EmployeeSupInfo.  Query for senator status
   in SqlSupervisorDao and set flag accordingly.
 - Prevent addition of a supervisor to their own SupervisorEmpGroup
 - Do not display senators in the employee select box in the employee
   select directive (used in attendance and accrual management pages)
   unless an attribute is set.

Fixes #10989