Fix name in "No time records need action" badge
On the Review Time Records page, "None of {Name}'s time records need
action." is displaying instead of "No time records need action", when
the default supervisor is selected and there are no overrides.  This
commit alters the creation of supervisor entries in the controller
such that the default supervisor is always flagged as the "fullEmpGroup"
and thus will display "No time records need action" when no 'SUBMITTED'
timesheets are present.