Modify Time Record cache behavior
Modified the time record cache to no longer cache time entries that were
initialized in the service.  These entries are not saved in the
database but are created to fill in entries for days that are not saved
in the database.  These cached entries have caused problems when there
was a change in a senate holiday and cached entries existed for the
modified holiday with/without holiday hours.

 - Modified cached record retrieval to return a copy of all cached
   records, instead of directly returning the cached record objects
 - Moved the initialize entries procedure to after cache
   storage/retrieval to ensure initialized entries are not cached
 - Modified Time Record copy constructor to deep copy time entries
 - Added performance tests for time record service tests to measure
   performance changes resulting from the above modifications.

Fixes #10976