Ignore blank temp. recs in attendance manage page
Temporary employees are not required to enter time each pay period, and
as a consequence, the records that are automatically generated for them
can build up.  All of these extra records can take up a lot of space on
the time record manage page, and are not of much interest to a
supervisor because the temp employee is not required to submit them.
This commit adds logic to ignore these records on the manage page.

Fixes #10890