Display 0 hours entered in time record detail view
The time record details viewer was displaying "--" when the employee
entered in 0 hours.  We would like 0 to be displayed in this case as it
indicates that a value was entered for the day (as opposed to a null for
no entry).  This commit adds a filter that implements this
functionality.

Fixes #11007