Fix filtering in employee select directive
In the employee select directive, the selected employee index is
defaulted to 0.  There is also a filter on the employee select dropdown
that filters out employees for various reasons.  If the first employee
(index 0) is filtered out, then they will still be selected by default
due to the default sel. index of 0.  This commit applies the filtering
when the employee list is set as opposed to when the list is displayed,
removing the possiblility that the default selected employee is filtered
out.

Fixes #11011