Modify "Active" column in grant display table
The table that displays privleges granted to a supervisor on the grants
page has a column called "Active".  This column displays the status of
the CDSTATUS column from the database.  The supervisor grants queried
from the database are filtered based on the CDSTATUS column and the
active column can be expected to always be true making it unnecessary.
This commit replaces the "Active" column with a "Status" column.  The
status column displays either "Pending", "Active" or "Expired" based on
the start and end dates of the grant relative to today.

Fixes #10922