Check for existing time record before insert
This adds a check in the time record dao that attempts to get the id of
an existing record with the same employee id and begin date before
inserting a new record.  This will prevent most instances where ESS
attempts to insert a duplicate time record, triggering a duplicate key
exception in the database.

Fixes #11023