Use employed dates in expected hrs projection
In the projection calculations, the expected biweekly hours (and by
extension the total expected hours) were being calculated based on the
total annual hours and the number of work days in the pay period.  This
does not factor in dates where an employee may have been unemployed.
This commit refactors the expected biweekly hours calculation to only
use dates where the employee was active.

Fixes #11042