Notify user if submitted record was modified
There are many time record submission errors that occur when a record is
modified by a third party (personnel/time record manager/second tab)
before the user submits it.  When a time record submission error is
received that could indicate that there was a third party modification,
the user is now prompted with a dialogue explaining the situation and
instructing them to refresh their time records and try submitting again.

Fixes #11030