Catch runtime ex. in time rec. scope validator
If a user submits a record where the record's previous state was in
the supervisor scope, the Permitted User Scope time record validator
will check to ensure that user is a supervisor of the employee for the
time period of the record.  If the user is not a supervisor, the
supervisor check will throw a supervisor exception, which is currently
uncaught and can result in a internal error sent to the user.  This
commit catches that exception and throws a time record error exception
instead.

Fixes #11051