Change expected hours computation strategy
The expected hours service computes expected hours up until the start of
a pay period.  This is good for calculating excess hours for time entry,
but not as useful when you are trying to look at expected hours for a
historical record.  This commit changes the expected hours service to
calculate expected hours including the given pay period.  The accruals
available calculation is also modified accordingly to still provide the
expected hours as of the last pay period.