Fix table header artifact on employee search page
On the employee search page, if the user opens the accrual history or
projections section, then opens another section above.  The accrual
table header will persist in its previous location on the screen.  This
is a known problem with the floating table header directive.  This
commit fixes this by triggering a reflow of the table header when a
toggle panel is toggled.

Refs #10693