Fix attend. and accrual history for senators
The attendance and accrual history directives currently error out when
attempting to display senator data.  This is because most senators do
not have any attendance or time records, and return an empty list for
the active record years api call.  This commit adds handling for empty
record years in both directives, displaying a message in this situation.

This commit also hides the accrual projection section for senators on
the employee search page.

Refs #10693