Fix accrual history year select on employee search page
The accrual history year select was not functioning on the employee
search page.  This was due to the fact that it was contained within an
ng-if element, moving it into a child scope from the parent controller.
This commit replaces the ng-if with an ng-show, putting the select back
in the controller scope, and causing changes in the select to register
in the controller.

Refs #10693