Prevent time entry id removal during split
Errors were occurring during time record splits in which there were
existing entries after the split date.  In these situations, the new
record generated in the split would attempt to insert a new entry on
top of the existing entry, triggering a duplicate key exception.

It would attempt to insert a new entry (as opposed to updating the
existing entry) because of the method 'ensureId' in the time record
dao.  This method would ensure that no time entry ids were included that
were not present in the existing record by setting all entry ids to the
values of the existing record.  This makes sense in the general case but
not in the split, where an entry is transferred from one record to
another.

This commit fixes the 'ensureId' method by preventing the id from being
set if the existing id is null.

Fixes #11096