Fix accrual computation for period end appoint
There were a number of issues with accrual computations that, when
combined, caused errors to be thrown for employees that were appointed
on the last day of a pay period.

The date from which accruals are calculated is normally the end date of
the last accrual record.  If no records exist, the employee's continuous
service date was used instead.  The problem with this is that the
accrual record end date is inclusive to the closed year, where the cont.
service date is the first day of the uncalulated period.  This commit
modifies the from date to the day before the cont. service date if no
annual accrual record exists.

When the initial accrual state is computed, pay type and min. total
annual hours are retrieved based on the "fromDate" described above.
The last value up to the from date was being used.  Now that the from
date can be the day before they were appointed, this will no longer
work, since the employee will have no values up to that day.  This
modifies that logic to use the FIRST value of the range of the from date
to the day after the from date.  This allows the employee's initial
values to be used if the from date is the day before appointment.

Fixes #11159