Enable accrual calculation when accruals disabled
In a 49a41f4567829316e3ff373a770b1a133568b984, I added a filter for
accrual calculations that disabled calculation for pay periods according
to the accrual allowed dates.  The accrual allowed dates are an
intersection of active employment dates, regular/special annual
employment dates, and dates where the employee is permitted to accrue by
personnel.

This was a mistake, as the intent was only to prevent calculation of
accruals for periods where the employee was not employed.  This has
caused problems for employees that are active and regular/special
annual, but not permitted accruals by personnel.  For these employees,
calculations should still occur, but accruals should never increment.

This commit narrows the scope of the pay period filter in the accrual
compute service to only exclude dates where the employee was not
employed.

Refs #11158
Fixes #11164