Restrict access to employee search page
Any user is currently able to access the employee search page, though
they are unable to view any user data due to access restrictions on the
API endpoints that provide that data.

This commit adds a permission check when the user attempts to retrieve
the employee search page template.  If they are not permitted to access
the employee search page, a template containing an error message is
returned instead.