Fix ordering of order columns in OrderBy
There was a bug in the OrderBy class where the order columns would not
be applied in the order specified in the OrderBy constructor.  This was
due to the fact that they were placed in an ImmutableSortedMap, which
ordered the columns based on their String value comparison.

This commit stores columns specified in the OrderBy constructor in an
ImmutableMap, which preserves insertion order.