Include inactive employees in employee search page
This commit adds the ability to search for inactive (in addition to
active, which were already searchable) to the employee search page.  Due
to the fact that the set of all past and present employees is an order
of magnitude larger than the set of active employees, a new approach to
search was implemented.

 - Add employee search query to employee dao (and employee service) that
   queries all employees in the database, matching by their first name,
   middle initial, and last name.
 - Expose employee search via REST api endpoint.
 - Modify employee search page to use employee search api instead of
   pulling in all active employees and searching using angular filters.
 - Use new third party angular module, ngInfiniteScroll, to paginate
   search results on search page.

Fixes #11189