Replace grid css in emergency notification page
For my original implementation of the emergency notification page, I
used css3 grid, which is relatively new as of this commit and will
probably never be supported by ie11.

I have replaced the grid styling on the page with a combination of
flex and simplegrid, a stylesheet that is already used on other pages.

Refs #11206