Restrict deactivation of timesheets on save
There is code in the saveRecord method in EssCachedTimeRecordService
that deactivates any previous, employee scoped records when a record is
saved.  This code is intended to clear out unused temporary pay records.
This code was being activated inappropriately when a supervisor approved
a record due to the non specific conditional that guarded the code.

This commit restricts the deactivation code in several ways:

- The deactivation code has been moved to its own method.
- The method is no longer reachable from the standard
  saveRecord(TimeRecord) method.  It has been moved to the
  saveRecord(TimeRecord, TimeRecordAction) method, which is used whenever
  a user saves, submits, or rejects a timesheet.
- The conditional that guards the deactivation method has been improved.
  The method is only called if the time sheet was submitted from an
  employee scoped status.
- A filter has been added to the deactivation method so that it will
  only deactivate records if they contain only temporary entries.

Refs #11231