Remove unused validation from time record manager
There was some unnecessary validation code in the time record manager
that would check the accruing flag of a time entry.  This flag is not
used nor is represented in the java model, and the result of this code
was not used in any way.

This code has been causing exceptions in the time record manager when it
attempts to get the accrual status of a temporary employee who has never
had an accrual status set.

This code is removed in this commit.

Refs #11238