Add resilient error handling for time rec. manager
There has been a consistent problem where there is a time record manager
exception for a single employee and it derails the entire time record
management process, affecting most of the employees in the senate if the
error happens early in the process.  Additionally, these errors will
generally happen several times before we know about them.  We usually
find out from an employee error report.

This commit addresses the first issue by compartmentalizing error
handling in the time record manager so that an error for a single
employee will not halt the process for the rest of the employees.  The
second issue is addressed by the addition of an alert email that is sent
to the administrator if any errors occur during the process.

- Add TimeRecordManagerError class to store info on time record manager
  exceptions.
- Add email service that will send a digest email containing info on
  multiple TimeRecordManagerErrors to the admin email.
- Catch and store time record manager exceptions for each employee and
  add to a running list of exceptions for the current run.
- Pass running list of exceptions to time record manager email service.

Refs #11238