Consolidate jackson config, fix for time objects
Some previous modifications to the jackson configuration removed a
deprecated module that enabled date/time objects to be serialized as iso
date/time strings.  This functionality was heaviliy ingrained in our
apis.  This commit adds the non-deprecated version of the module
returning the api response formats to normal.

Additionally, this commit consolidates the two sources of Jackson object
mappers: the static OutputUtils and spring bean provider JacksonConfig.
JacksonConfig now provides the same Jackson object mappers that are
configured in OutputUtils.