Add built in methods for common trans.history uses
One of the primary uses of the TransactionHistory object is to get a set
of dates for which a certain attribute holds true for an employee.  This
is somewhat difficult because the TransactionHistory provides a tree map
of effective dates to values, instead of what we want, which is a
RangeSet.  The process of converting a tree map to a range set takes at
least 5 lines of code and is repeated throughout the Time codebase.

This commit adds a function that further abstracts the behavior of
converting a tree map to a range set. It also adds several
range-set-returning functions to TransactionHistory for common uses.

- Replace existing RangeSet getEffectiveRanges functions with one that
  takes a sorted map and a predicate.  This fuction converts the sorted
  map into a range map, filters the range map based on the predicate,
  and returns the remaining keys as a range set.
- Add range set functions for several transaction fields in
  TransactionHistory.
- Replace existing tree set -> range map code with new methods from
  TransactionHistory.