Add a custom view for the supply SFMS sync process
Adds a new requisition view to be used specifically for serializing
requisitions into xml for the supply SFMS sync process. This will
prevent potential bugs in the synch process caused by changes to view
objects.

Also fixes a bug introduced in commit 1b9fcb5. That commit added the
requirement for a LineItems tag in the XML sent to SFMS. However, there
is an edge case where supply would correctly try to sync a requisition
with no line items, causing an inaccurate exception. This commit
addresses this by not sending these requisitions to SFMS.