Fix extended supervisor permissions
We recently made a modification to allow Senators to view and approve
time records for all employees in their department, regardless of their
depth in the supervisor tree.  This modification was incomplete, because
senators would receive an authorization error when trying to view the
approval screen for employees more than 1 level down the tree.

This commit modifies the supervisor authorization service to grant
supervisor specific permissions to all supervisor's in a Senator's tree,
instead of just those directly under the senator.

Fixes #11245