Modify mobile subscription in emerg. alert model
Previously, the emergency alert model contained a boolean field called
"smsSubscribed" for storing whether or not texts were sent to the alert
subscriber.  This field required the user to enter a mobile number,
which would receive calls regardless of whether the sms subscription
flag was set.

This commit replaces that field with a more granular mobile options
field that permits the user to subscribe to only calls, only texts, or
both.

- Added MobileContactOptions enum.
- Replaced smsSubscribed field with MobileContactOptions in
  EmergencyNotificationInfo and related view.
- Added patch script to convert emergency notification info table to use
  MobileContactOptions enum instead of sms_subscribed flag.
- Modified emergency notification info dao to use new contact options
  enum.