Allow accrual comp. when accrual is forbidden
A previous commit restricted the computation of accruals for periods
where an employee was not accruing (not employed, temporary, accrual
flag = false etc...).  This caused some problems, one of which was that
annual employees that were not permitted to accrue could not enter time
due to the requirement of the time entry page to display accruals.

This commit rolls back the previous changes, and going forward we will
find a new way of restricting accruals that handles all of these
different cases.

Fixes #11248