Fix bug with alert info validation
On the alert info page, a user is not allowed to submit alert info if
there is a mobile number and invalid mobile options (neither calls or
texts are selected).  If there is no mobile number then the mobile
options do not matter, but if invalid mobile options are posted, it
causes an error on the backend.

This commit will reset the mobile options to both calls and texts if
neither is selected and no mobile number is present when the user
submits the form.

Also, an unused function for toggling mobile options is removed.