Adds IP address authentication for API access
Clients with an IP address matching the new property
'auth.api.ip.whitelist' are allowed to access API endpoints without
specifying a username and password and are given admin permissions.

If this client subsequently tries to access non API endpoints, they will
be logged out and redirected to the login page so they can do a
traditional login. This is necessary so information that is required by
the UI, like employee data, gets initialized properly and attached to
the subject.