Allow editing of requisition notes accross revisions
Previously, a note on a requisition was set on a single revision. It
could not be edited and could only been seen in the detailed requisition
history page.

This change makes the note field act like a single note for the entire
requisition instead of a note for a single revision. The note can now be
seen and edited at any time through the modals on the fulfillment page.