Improve front end alert email validation
The front end email verification (via html 5 email input) is not as
strict as the backend verification (apache commons-validator).  This
causes the potential for the user to trigger an internal error by
entering in an email that passes front end validation but fails back end
validation.  This commit improves front end validation and handles the
scenario where the back end validation rejects an email address.

- Add enum for all possible alert error codes
- include alert error code enum in alert exceptions/alert error response
- If the invalid email error code is detected on the front end, show an
  error modal that identifies the invalid email address.
- enhance front end email validation by requiring a top level domain in
  the host portion of the email.