Remove allowance payment filter
A filter in the service that calculates allowances, filters out payments
from the calculation that do not fall in the range of unpaid dates.
As payments are received, the range of the payment is subtracted from
the unpaid date range.  This was causing an issue where a payment was
filtered out because it was for dates that were contained by the date
range of a previous payment.  This filter is somewhat redundant because
the payments are already restricted to the current year in the query.
It is removed in this commit.

Fixes #11302