Fix accrual computation for new employee
When getting the available accruals for a pay period, the accrual
compute service tries to get accruals for that pay period and the
previous pay period.  When it attempts to get the accruals for an
employee's first pay period, it looks for the accruals of the previous
period where the employee wasn't working.  This was causing a problem in
the accrual compute service, which would throw an exception if the first
pay period requested was not a valid accruing period for the employee.

This commit prunes the queried pay periods before making this check.

Fixes #11301