Retrieve override info in sup emp group service
There was an issue on the approval page occurring when there was a
supervisor override on a supervisor that was not in the override
grantee's extended supervisor employee group.  This error occurred
because all names on the page are retrieved from a map generated by the
ext. sup. emp group. This commit fixes that issue by retrieving
additional information in the angular sup emp group service.

 - Add a method to retrieve an employee's supervisor overrides in
   sup-emp-group-service.js.
 - Call override retrieving method in sup emp group service
   initialization.
 - Add the names of all supervisor overrides retrieved to the sup emp
   group service name map.

Fixes #11321