Merge connected employee sup info
There was an issue with an employee who was terminated and rehired the
next day.  Because there was no discontinuation of his employment, he
did not have a split timesheet.  His employee info for the supervisor,
on the other hand saw this as two distinct periods of employment.  The
permissioning system for time records requires that a permission enclose
the date range of the time record in order to allow approval.  Because,
there were to distinct periods of supervision, there were two distinct
permissions, neither of which enclosed the time record range.

This commit modifies the time record dao to consolidate EmployeeSupInfo
that have connected date ranges.  This joins the two periods of
employment and allows the permission to enclose the time record range.