Add warning logs for common API exceptions
The majority of the currently handled api exceptions should not be
occurring.  If they do occur then it typically indicates an issue with
the front-end or a data issue.  The result of these exceptions are sent
back to the requestor via an error response and the front end displays
an error pop up with the option of sending the error as an error report.
The users do not always send the error report, and they will often only
send it after getting several errors.  We need to be able to track these
errors without relying on users to click the error report button.

This commit adds warning logs for common api exceptions, allowing us to
maintain a complete record of errors in the log.