Increase permissions for sup. override employees
Employees that have been granted a supervisor override need to be able
to act as the supervisor for the duration of the override.  This means
that they can view and approve records of override employees for time
periods before the override began.  The current permissioning scheme for
override employees only grants permission for the period of the
override.

This commit removes the restriction of the override duration from the
range of permissions granted.

Fixes #11403