Improve pre-sfms emp. handling in supervisor query
Senate employees that were hired before SFMS was built do not have
appoint transactions.  For these employees, we treat the very first
personnel and payroll transactions as the appoint documents for
personnel and payroll values respectively.  The supervisor query (which
looks for personnel values), was not getting the very first personnel
value.  It was sorting by effect date only.  The proper way is to sort
by effect date and transaction creation date because there can be many
transactions with the same effect date.  This was causing some null
pointer exceptions as a side effect because the wrong transactions were
being used (with incomplete values).

This commit factors in transaction creation date when determining the
appoint transaction for pre-sfms employees.