Fix missing Supervisor employees.
When there are multiple continuous records with the same
Supervisor/Employee combination, the records are merged. When the later
record had a null Supervisor End Date and the earlier record had an actual
date, the merge was taking the earlier date. It should have taken the null
value (supervisor is still the current supervisor). This fix will compare
the null date as being later and take that over an actual End Date.

Fixes #11513