Fix bug with empty ranges in transaction history
The TransactionHistory method getEffectiveEntriesDuring fails when given
an empty range.  This happens because the method uses TreeMap operations
which do not support empty ranges.

This commit modifies the getEffectiveEntriesDuring method so that it
returns only the initial entry or an empty map if no initial entry
exists, when an empty range is passed in.