Handle partially appointed emps in supervisor dao
Employees info is typically generated in two stages: payroll data and
personnel data.  There can be a gap between these stages, creating a
time period where the employee has null values for key data points.
This is already handled in the transaction history, but is a little
rough in the supervisor dao, which generates a lot of exceptions when
this happens.

This commit adds more robust null checking for these employees.  If any
of the required values are null, a single line warning message will be
printed, and the record will be ignored.

Fixes #11118