Prevent supervisor override loops
A relatively common scenario of a supervisor override would be when a
trusted employee is granted the permissions of their supervisor.  In the
current state of ess, this would also grant the employee supervisor
permissions over theirself.  An employee should never to approve their
own timesheet.

This commit adds a check in the addSupOverrideEmployee method in
SupervisorEmpGroup that aborts any attempt to add an override employee
with the same id as the supervisor.

Fixes #11408