Fix expected hours calculation
The method that was calculating expected hours did not reflect the
standard procedure.  According to senate rules, you first need to
calculate and round the hours an employee is expected to work each day.
Then this daily value is multiplied by the number of days the employee
is expected to work.  The existing method in ess did the same
calculation, but did not round until after the the days were multiplied.

This commit adjusts the method to follow the standard way of calculating
expected hours.

Fixes #11572