Fix bug in supervisor emp group
I flattened the storage of override employess in SupervisorEmpGroup in a
previous commit.  This was foolish.  I thought that looking up override
employees by id was never used, but it was, in the method
hasEmployeeDuringRange.  In my modification of this method, I changed it
from searching through emp. sup. infos of the given employee, to
searching it for all employees.  As a result, this method will always
return true if there are any override employees active during the given
range.

To fix this I am reverting to storing override employees in multimaps.
The storage of supervisor overrides is still different from what it
originally was, because its original storage data structure (a table)
did not allow for multiple spans of employment for a single employee.

Fixes #11576