Add debug redirect option for ess emails
This commit adds a test mode option for ess emails in app.properties.
If this flag is enabled, all emails will be sent to the configured test
mode address rather than the usual recipient.  If this option is
enabled, the address of the recipients that would have normally received
the message are appended to the message subject.  The option is enabled
by default and will prevent startup unless it is explicitly disabled, or
a valid test mode email address is set in the properties.

I have packaged this feature in with modifications to existing code that
appends the runtime level to email subjects when the runtime level is
not prod.  This code is now located in an override of
JavaMailSenderImpl#doSend, which is called any time any email is sent.

Additionally, I have added an enum for runtime level, which is used in
this service along with the CommonAttributeFilter.