Fix expected hours calculation.
It turns out that the expected hours calculation is wrong again.  In the
most recent calculation, we would calculate and round the hours an
employee would work each day based on the minimum total hours value.  I
have learned that the proper way is to first calculate the number of
hours worked each day, multiply by 10 and THEN round to get the number
of hours worked per pay period.  Then, for calculating expected hours
over an interval, we divide the pay period hours by 10, multiply by
number of days in the interval, then round to nearest .25.

This commit implements the correct* expected hours formula.
* as far as I know at this moment

Fixes #11588