Fix bug in expected hours calculation
In a previous commit I consolidated the two methods responsible for
calculating expected hours: one that takes a pay period returning a
number, and another that takes a date range returning an ExpectedHours
object.  I did not realize that the number returned from the first
calculation was the number of hours expected after the end of the pay
period instead of the expected hours prior to the pay period.  This
caused all expected hours calculations to be off by 1 pay period's worth
of hours.

In this commit I address this issue by further standardizing the
expected hours methods to both return ExpectedHours objects.  This way,
you know what you're getting from the method, and can extract the
desired value from the object.  I also fixed a related bug in the
AccrualComputeService where a date that was labeled as a "from" date was
actually the "to" date of the previous pay period.

Fixes #11572, #11588