Fix bug with pay period number.
In the recent past, I added a method to the PayPeriod class that adds a
letter designator to the end of the split pay period at the start/end of
a year.  This allowed the pay period number to be unique within a year.
I did this by modifying the getter method of the pay period to add the
letter if the pay period was split.  This has caused problems when the
pay period is serialized and deserialized from its json value.  When
deserialized, the augmented pay period number is set for the new pay
period.  This causes it to differ from the same pay period that was
just read from the database.  This bug has resulted in validation errors
when employees try to save/submit time records for the last pay period
of this year (a split period).

This commit fixes this issue by stripping any non-numeric characters
when the pay period number is set on the PayPeriod object.  This way,
the pay period number is stored in its original form and will not be
different if the pay period is serialized then deserialized.

Fixes #11594