Show message for inactive emp. reminders
Supervisors can still see and approve unsubmitted records for employees
that have been terminated, but these employees cannot log in or receive
email because they have had their senate accounts revoked.  If the
supervisor were to attempt to send an email to them, an exception would
be raised.

This commit addresses this situation by displaying a message to the
supervisor if they attempt to send an email reminder to a terminated
employee.  This message will inform them which employees are inactive
and instruct them to call personnel, who will instruct them on getting a
paper timesheet for the employee.

Fixes #11294