Change approach for policy links
This commit modifies the policy model so that Policy is just storing the
filename, not the entire link.  PolicyView will have the full link, and
will generate the full link in the constructor using a context path
passed in from the controller and a predefined policy directory.

Also added a dummy policy in the newly created policy directory.