Require full scroll to acnowledge document.
Modify the acknowledged document view so that it requires the user to
scroll through the entire document to submit an acknowledgement.

There are some hacks involved due to the nature of the pdf viewer.  In
order to detect scrolling of the document, the size of the pdf document
is calculated and the embed tag is expanded to fit the entire document.
The expanded embed tag is placed inside a scrollable container which is
monitored via a directive.